Move error logging into the Analyzer
Now that the Analyzer needs a Logger and has a concept of "initial"
analysis, this seems simpler.
