+++ linker/shared/src/main/scala/org/scalajs/linker/frontend/BaseLinker.scala
@@ -13,7 +13,6 @@
 package org.scalajs.linker.frontend
 
 import scala.concurrent._
-import scala.util.Try
 
 import org.scalajs.logging._
 
@@ -38,7 +37,7 @@ final class BaseLinker(config: CommonPhaseConfig, checkIR: Boolean) {
 
   private val irLoader = new FileIRLoader
   private val analyzer =
-    new Analyzer(config, initial = true, checkIR = checkIR, irLoader)
+    new Analyzer(config, initial = true, checkIR = checkIR, failOnError = true, irLoader)
   private val methodSynthesizer = new MethodSynthesizer(irLoader)
 
   def link(irInput: Seq[IRFile],
@@ -49,7 +48,7 @@ final class BaseLinker(config: CommonPhaseConfig, checkIR: Boolean) {
     val result = for {
       _ <- irLoader.update(irInput)
       analysis <- logger.timeFuture("Linker: Compute reachability") {
-        analyze(moduleInitializers, symbolRequirements, logger)
+        analyzer.computeReachability(moduleInitializers, symbolRequirements, logger)
       }
       linkResult <- logger.timeFuture("Linker: Assemble LinkedClasses") {
         assemble(moduleInitializers, analysis)
@@ -70,39 +69,6 @@ final class BaseLinker(config: CommonPhaseConfig, checkIR: Boolean) {
 
     result.andThen { case _ =>
       irLoader.cleanAfterRun()
-    }
-  }
-
-  private def analyze(moduleInitializers: Seq[ModuleInitializer],
-      symbolRequirements: SymbolRequirement, logger: Logger)(
-      implicit ec: ExecutionContext): Future[Analysis] = {
-    def reportErrors(errors: scala.collection.Seq[Analysis.Error]) = {
-      require(errors.nonEmpty)
-
-      val maxDisplayErrors = {
-        val propName = "org.scalajs.linker.maxlinkingerrors"
-        Try(System.getProperty(propName, "20").toInt).getOrElse(20).max(1)
-      }
-
-      errors
-        .take(maxDisplayErrors)
-        .foreach(logError(_, logger, Level.Error))
-
-      val skipped = errors.size - maxDisplayErrors
-      if (skipped > 0)
-        logger.log(Level.Error, s"Not showing $skipped more linking errors")
-
-      throw new LinkingException("There were linking errors")
-    }
-
-    for {
-      analysis <- analyzer.computeReachability(moduleInitializers, symbolRequirements, logger)
-    } yield {
-      if (analysis.errors.nonEmpty) {
-        reportErrors(analysis.errors)
-      }
-
-      analysis
     }
   }
